Moved types to own file and created Testimonial type
